Replace or_ with in_ function for searching queries
Now in several queries we use or_ to define filter conditions.
But it's known that in_ works faster [1], especially when filtering
occurs on non-indexed fields. Therefore it's suggested to replace
or_ with in_.

Local tests showed that the performance for such requests increases
to 30%, depending on the size of the database.

[1] https://stackoverflow.com/questions/782915/mysql-or-vs-in-performance

Change-Id: I47caf99b9b1a66713f9d92bca0d5f692b7f2a717