Fix inconsistencies when setting policy values
This patch fixes inconsistencies between two ways of setting policy
values: as a variable and directly in workflow definition as a constant.

Inconsistency #1:
For policies 'wait-before', 'wait-after', 'retry', 'timeout', 'concurrency'
there is a difference on how they are executed if value is 0.
If the value is hardcoded in workflow, the policies are omitted [1],
but if a user defines them as a variable, then the policies are applied.

Inconsistency #2:
Policy values in workflow definitions cannot be negative numbers (validated
by schema) [2], but if a user sets them as variables it's okay [3]. It
happens because the schemas are different for both cases.

[1] https://github.com/openstack/mistral/blob/master/mistral/engine/policies.py#L83
[2] https://github.com/openstack/mistral/blob/master/mistral/lang/v2/policies.py#L27
[3] https://github.com/openstack/mistral/blob/master/mistral/engine/policies.py#L161

Change-Id: I660ec2fe00e9f524292957560548447e517332fc
Closes-bug: #1731100