Fix sporadically overwriting of finished workflow execution state
Sometimes it happens when Mistral rewrites the state of an
execution that has already been finished.

It happens because of possible race conditions in mistral when a task turns
the state and finishes the execution, but a scheduled call is not awere about
that and changes the state again.

To fix the bug we use WHERE clause during the execution update.

Change-Id: I96c16845e6cee08def9e50cfaeac8f36940f96dd
Closes-bug: #1731418