Cache action definitions
Now to perform some action mistral gets its definition from
the database first. It's not really optimal, because if there are
a lot of similar action calls, mistral will reread the same data
from db. It increases the whole execution time and the load on the
database.

To improve the performance it's suggested to cache read definitions
and take them from the cache instead of the database in the subsequent
times.

Cache ttl can be configured with ``action_definition_cache_time``
option from [engine] group. The default value is 60 seconds.

Change-Id: I330b7cde982821d4f0a06cdd2954499ac0b7be37