Make more CI jobs voting
A recent regression showed up in the tempest tests, but as they were not
voting it was ignored and not spotted until later. This caused lots of
time to be wasted debugging.

We should make more of the jobs voting and fix any instability (or
remove particular tests that are unstable).

Related-Bug: #1736685
Change-Id: Ifd6f1d659db06bca3e6417b9235258fd02f1da88