Remove the invalid toctree
The reference to v2/index doesn't exist, causing the build to fail.
Given the api-ref documentation is currently empty, removing the toctree
wont cause any issues.

This makes the build green (test with `tox -e api-ref`), which will allow CI to
be enabled which may help further work on the api-ref.

Change-Id: I9893c3f629255ae2ae879a60fa341e3775b426ed