Really make the cron trigger execution interval configurable
Note: This is an update to I9060253bc416be28af4ef81f3edf694059d92066,
      which almost worked but due to a race condition with config loading
      the default value may have been used. This delays the decorator
      execution to allow the config to be fully loaded.

The cron_trigger subsystem in Mistral queries the database every second
to look for triggers that require execution. This can be very wasteful
if your deployment only has cron triggers that run infrequently (every
hour or day etc.). Letting operators configure this value reduces the
load and allows the cron triggers to be useful in more scenarios.

Operators do need to be aware that this configuration would limit the
frequency of execution. For example, if the value is set to 600 then cron
triggers configured to run every minute will only run every 10 minutes.

Related-Bug: #1747386
Change-Id: Ia12dcf7fa6b1784be226d223eb89b114c8e4b4f5