Use eventlet-aware threading events
For the gory details behind this, see:

https://bugzilla.redhat.com/show_bug.cgi?id=1384183

In summary, if you use threading.Event on python2 with eventlet, it
unnecessarily spams system calls to epoll_wait and burns a *lot* of
CPU cycles by constantly context-switching.

I added eventletutils.Event to oslo_utils a while back to
intelligently use the eventlet Event object when eventlet is being
used, and to fall back to the threading one otherwise.  So just use
that and it will do the right thing.

Depends-On: https://review.openstack.org/#/c/558879/
Change-Id: I6ca0d0fc2a2ff393bab502a5f705fb0d63c32c6b