Fix join examples and text
The 'join' example for using a number did not have a proper workflow.
It looks like it was copied from the 'all' example and not completely
changed. Also, disambiguate the wording of the description for that
example. It used the word 'once' in a place where it looked as though
it meant 'one time' instead of 'as soon as'.

Change-Id: I9b89768c0f0b7d325db08ea9345965a9263a64c8
Closes-Bug: #1761827