Make cron-triggers not play catchup
Currently if the API is stopped for a length and restarted the crontrigger
jobs will have their next execution in the past and will do a catchup
running a lot of times. That happens because the next execution is
calculated based on the first time it was executed. This change makes it be
the next execution time be the max of the calculated time and utcnow so
that if the server is lagging we keep uptodate with the calculated next
execution times.

Change-Id: I05b35261b1adbe17a1b0316eee95da12634c3978
Closes-bug: 1719882