Return the result of the MistralHTTPAction
If the HTTP request fails, we need to fail the task. Returning the error
from the parent class will do this. While this means we also return the
success result it will be ignored by the Mistral engine.

Credit to @lijianying10 on GitHub for sending this fix via a pull
request. Tests were then added to verify the change.

Closes-Bug: #1779973
Change-Id: Ib8754c8de2d6677d71383b3793d0fa168be575f5