Allow engine commands as task name
A change for disabling some task names has introduced a massively
backward incompatible behavior. E.g. even though there is a 'noop' engine
command, the usual way of handling noop is in many cases is to create a
task called the same. The other commands are not used that often but
noop is massively present in currently deployed workflows and it is
not possible to mitigate the error if the workflows are coming from
3rd parties.

This change re-enables the usage of the engine commands as task names
and adds documentation on why this is a useful feature.

Change-Id: If90ee5f787e4587a25c156d12c7750407081bf0d
Related-Change: https://review.openstack.org/#/c/535297