New rerun events
Currently, all workflow actions, for example, start, pause, resume,
stop have own events except rerun. Even when we rerun a task, event
"TASK_LAUNCHED" doesn't propagate to a parent workflow and task.

Add new WORKFLOW_RERUN and TASK_RERUN events. These events propagate
to a parent workflow and task. You can find a example of new events in
the following tests:
* test_notify_rerun_task
* test_notify_rerun_nested_workflow

The recursive parameter was removed in the workflow set_state method
because it only is used in the rerun method. Recursive rerun was
extracted in the separate function.

Change-Id: Ib8bd3cf5c314ab4529260736cc0b7d792ce53171
Closes-Bug: #1790441
Signed-off-by: Vitalii Solodilov <mcdkr@yandex.ru>
Co-Authored-By: Oleg Ovcharuk <vgvoleg@gmail.com>