Docs improvements: task timeout, global context, Docker and jinja
A not obvious point for users is the task does not retry after a
timeout is triggered. Added clarification in the retry section.

The documentation contains enough example with Jinja usage. Improved
only the create_vm workflow definition.

Added global publishing to the doc. It is brash copy-paste from
https://specs.openstack.org/openstack/mistral-specs/specs/pike/approved/advanced_publishing.html without mention of atomic publish.

Move Docker guides to the installation section.

Change-Id: I149b2e1dff7f86bd356f4dd2f758659469e6a4a8
Closes-Bug: #1789384
Closes-Bug: #1690156
Closes-Bug: #1779244
Signed-off-by: Vitalii Solodilov <mcdkr@yandex.ru>