An execution hangs in the RUNNING state after rerun
When we rerun an execution we must create the "_check_and_complete"
delayed calls for all parent workflows. The problem was that we created
the delayed call only for the rerun execution and its parent.

Recursive rerun was extracted in the separated function. Because we
need to execute some additional operations, for example, create delayed
call for every a rerun execution.

Closes-Bug: #1792451
Signed-off-by: Vitalii Solodilov <mcdkr@yandex.ru>
(cherry-picked from commit 041f3bd35c4f438c27087f5b48784055ae75f1c8)

Change-Id: I64de0c09bbe63401edead2f3ada5859b1db4a771