Update version.version_string to actually be a string
pbr provides VersionInfo.version_string() as a method to determine
the version.  mistral.version.version_string can hide this and
provide a uniform string interface to config and launch.

This fixes a problem with mistral-server --version generating
an exception because the version passed to argparse was a method
instead of a string.

Change-Id: Ie468685e4360bfaec5d82b02f8cf1a27a93bcd94
Closes-Bug: 1796921