Replace openstack.org git:// URLs with https://
This is a mechanically generated change to replace openstack.org
git:// URLs with https:// equivalents.

This is in aid of a planned future move of the git hosting
infrastructure to a self-hosted instance of gitea (https://gitea.io),
which does not support the git wire protocol at this stage.

This update should result in no functional change.

For more information see the thread at

http://lists.openstack.org/pipermail/openstack-discuss/2019-March/003825.html

Change-Id: Ia4da4f94ecada34246a4db95174fa42595f9ff86