Fix adhoc action lookup
Now to find an adhoc action definition by name action hanler takes
the name in runtime context first, and then provides this value along
with workflow name and workflow spec name to resolve_action_definition
method.

This method tries to find a definition by full name, i.e.
"wb_name.action_name" if wf_name != wf_spec_name, and, if nothing was
found, looks for a definition just by name, i.e. "action_name".

There is an issue, because runtime context already contains full name
for the action and substitution of a workbook name as a prefix is not
necessary.

It leads to the fact, that for all adhoc actions mistral tries to find
them by "my_wb_name.my_wb_name.my_action_name" first, and then, when it
fails, it looks for "my_wb_name.my_action_name".

This results in double the number of queries to the database and highly
innefficient.

This commit solves the issue by not providing workflow name and workflow
spec name to resolve_action_definition method for adhoc actions.
After that, the method will search for the action immediately by its
name.
For system actions this substitution is not necessary there either, so
the code can be safely deleted as well.

Change-Id: I3b880ef29057afb041edd9095c279a944c77f2dd