Rework finding indirectly affected created joins
Current algorithm of finding the joins requires a recursive search
through the execution graph, which leads to a large number of calls
to the database.

To optimize it we introduce a new algorithm that requires only one db
request. It downloads all potencial join task ids and names, and then
analyzes them without any additional db calls.

Change-Id: Ic73f2112406e681ae8a2aa67bcbccebe488fc03c