Limit max search depth
Now, to find an induced join state for a task we prepare a set of
all task's parent names and then perform a db request to fetch all
the tasks in one request.

But sometimes the depth of the search can be great, and even searching
of parent task names can be time-consuming.
To prevent this situation this patch introduces limited depth search,
Which finds parent tasks at a limited depth first, and loads the rest
only if necessary.

Change-Id: I7126b7c9652f190ec5c8423fad60d18729840a3a