Remove unneeded Zuul branch matcher
We have implicit branch matchers, so there's no need to add a check for
not-ocata etc, this job is only run for the branch it's on - like master
now.

Remove it to not confuse Zuul when multiple branches matches and the job
is different.

Leave the master lines in, those restrict it from not running after
stable creation.

Change-Id: I22ae60149e2539eab4b8b39b924ba49ba51ab3d6