Mask sensitive data when logging action results
When passwords or other sensible data is returned by a Action, they can
be logged by Mistral. This change uses the password masking
functionality used in mistral-lib and privided by oslo-utils.

This function uses the standard cut_repr method in mistral-lib, which
also means the output is more standardised.

Related-Bug: #1850843
Change-Id: I01bf47f7a83102a1a16b15bf0bbb4021707e11fe