Make the mistral devstack jobs non-voting in Rocky
These jobs are problematic when trying to merge a security fix into
Rocky[1]. This change has passed and merged intro master, Train and
Stein. It even passed CI once in Rocky (on Nov 19th) but then failed
in the gate.

This change disables CI to allow this security fix to land and will
be reverted afterwards where we can investigate the unstable CI.

[1]: https://review.opendev.org/#/c/694733/

Change-Id: I26d587241f765b13e135be5c4cf90976903af9ea