Remove the TripleO standalone job
Despite the comment in .zuul.yaml this job actually doesn't use Mistral.
It is installed, but not used. This can be seen in the Mistral logs
which have minimal output (other than that of the workbooks being
loaded).

https://c867b5a19c86ae8c3ab7-a51f273e08e05c861a2fa6ace8476176.ssl.cf1.rackcdn.com/695212/2/check/tripleo-ci-centos-7-scenario003-standalone/4827980/logs/undercloud/var/log/containers/mistral/

Th other TripleO job also covers this case and then uses Mistral. This
patch removes the job which has been wasting resources for quite some
time now.

Change-Id: Ie3381c7226d50591192ae61697bdb0d134ddf4f1