Make mistral reproducible
Whilst working on the Reproducible Builds effort [0] we noticed that
mistral could not be built reproducibly.

This is because some help output enumerated the constants in a Python
set without a sort which does not have an implicitly deterministic
ordering. Patch attached that applies a sort to the list.

 [0] https://reproducible-builds.org/

Note this was reported in Debian at:
https://bugs.debian.org/960591

and was fixed in the Debian package.

Change-Id: Icb899483021777724462e926732a182f2a3390b0