Fix compatibility with oslo.context >= 4.0.0
The tenant argument of RequestContext is no longer available since
oslo.context >= 4.0.0 . This change fixes the compatibility issue
caused by that removal.

Note that this still keeps reference to 'tenant' argument to make
the code compatible with older oslo.context, but that can be removed
once oslo.context >= 4.0.0 becomes available in upper-constraints.

Change-Id: Ie671f50e5ff5a7c746f9e95691eaf4dd19937b52