Remove extra description on cron trigger executions
When creating a periodic workflow execution from a cron trigger, the
description field was filled with info like cron trigger ID, name, etc.

The cron trigger name can be set by user and is then injected in the
description field.
Beside the fact that this is not useful, it may also be dangerous.

The purpose of this patch is to remove the name from the description so
that the lenght of description field is not user dependant anymore.

Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>
Change-Id: I326400a4a04fba71ccc90545909212a41d6b3b13