Disable mysql and postgresql unit tests
The purpose of units is to tests input/output of different functions /
methods of a program.
This is not to test the fact that it works using mysql or any other sql
engine.

So, lets focus on units that are easy to run locally with sqlite (while
mysql / postgresql needs specific installation)

Change-Id: I7d7c499c7176cbd75f02b2aff5d070733305c4cd
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>