Dynamically generate tests for default tasks
CI is flapping from time to time because the test_task_defaults was
taking too long.
Usually, this test is taking between 50 to 80 secs, but it could take
more than 160s, which is a hard timeout for us.
Let's split those tests into multiples small unitary tests (dynamically
generated) so that we never trigger this timeout.

Change-Id: I9d23b541d48765ad25d748e356e2013d90f940b6
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>