Fix join test_triggered_by_error
The test test_triggered_by_error is failing with new eventlet release.
Here is why.
When a task fail, the worflow is set to error.
Beeing in error prevent any further task to be executed.

Because of uncontrolled behavior, this test was expecting that the join
task was triggered even if a previous task was in error.
This is not always true.

The fix is to make sure the join task is triggered on error by using
on-complete instead of on-sucess.

Change-Id: I9224d53a171dc7a9f46e51cc2c6f43d2ea7852c5
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>