Deprecate mistral coordination
Mistral uses tooz coordination only for /v2/services endpoint, which is
giving the list of mistral services running per workers.
This is not enabled by default and need a backend_url to be set in
config for this feature to be enabled.

During the eventlet removal, we identified that some drivers in tooz
rely on the fact eventlet monkeypatching is enabled.

Since we want to get rid of eventlet, and because we think this services
endpoint is not widely used, let's deprecate that feature so we can move
forward with eventlet removal.

All coordination code may be deleted during the next cycle.

Related-bug: #2099954
Change-Id: Ia1a27d196b03297e25388b22181cc37f74837fa8
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>