Do not call setup_db from engine
Calling setup_db is creating the db structure if missing (because it's
calling models.Workbook.metadata.create_all after).

This is not what we want since we want the operator to manage the
creation of the db using mistral-db-manage like any other regular
openstack service.

I suspect that this call was mostly useful for tests.
This is a real issue in production because it create the DB structure
without using the correct alembic migration path (and thus,
alembic_version table is missing).

Change-Id: I51f4128dd023846a4406cd1b414a56d741035483
Signed-off-by: Arnaud M <arnaud.morin@gmail.com>